ledchaincomm: fix power calculation regression (caused by adding 16bit PWM support)
while 32bit lightPowerPWM would wrap around only after 16384 RGBW LEDs fully powered (beyond our scope),
lightPowerMw could wrap around already at >546 WS2815 or >257 WS2813 when doing calculation in 32bit
because of mW factored in (85..120).

So casting calculation of PWM*millwatts to uint64_t, then dividing by PWMMAX will keep result in 32bit range (4kW...)

+++ ledchaincomm.cpp
@@ -1338,7 +1338,7 @@ MLMicroSeconds LEDChainArrangement::updateDisplay()
               }
               // end of one chain
               // - update actual power (according to chip type)
-              lightPowerMw += lightPowerPWM*chip.rgbChannelMw/PWMMAX + lightPowerPWMWhite*chip.whiteChannelMw/PWMMAX;
+              lightPowerMw += (uint64_t)lightPowerPWM*chip.rgbChannelMw/PWMMAX + (uint64_t)lightPowerPWMWhite*chip.whiteChannelMw/PWMMAX;
             }
             // update stats (including idle power)
             mActualLightPowerMw = lightPowerMw+idlePowerMw; // what we measured in this pass
